Add files via upload
The first Jordan digital currency created in 2023 by developer Abdullah Sobh. The currency is based on investing in stock markets, designing and selling NFT boards, which returns this business at a profit, and this profit is divided into two parts. The first is to inject liquidity into the currency and create projects that develop from the currency, which makes the currency advance, and this is what is required, and the second is the distribution of profits to currency holders, and it will be a currency distributed for annual profits only to holders of the JD digital currency, and for all new things. Followus on social media channels



 EMIL SUPPORT

SUPPORT@JOCOIN.WEBSITE